requirements.txt
pytest == 8.3.3
allure-pytest == 2.13.5
requests == 2.32.3
faker == 30.0.0